llm: move auth back to main crate
meant to push this into
https://github.com/agentgateway/agentgateway/pull/2469 but missed it
