+++ crates/agentgateway/src/http/auth/azure.rs
@@ -1,7 +1,6 @@
 use std::sync::Arc;
 use std::sync::atomic::{AtomicUsize, Ordering};
 
-pub use agent_llm::auth::AzureCredentialCache;
 use azure_core::credentials::{AccessToken, TokenCredential, TokenRequestOptions};
 use azure_identity::UserAssignedId;
 use secrecy::{ExposeSecret, SecretString};
@@ -36,6 +35,24 @@ pub enum AzureUserAssignedIdentity {
 	ClientId(String),
 	ObjectId(String),
 	ResourceId(String),
+}
+
+/// Per-instance credential cache for [`AzureAuth`].
+///
+/// Each [`AzureAuth`] value owns its own cache so that different backends
+/// (e.g. two `ExplicitConfig` entries with different client secrets) get
+/// independent credentials instead of sharing a single global cache.
+/// Clones share the same underlying `Arc`, so the credential is built at
+/// most once per config instance.
+#[derive(Default, Clone)]
+pub struct AzureCredentialCache(
+	Arc<tokio::sync::OnceCell<Arc<dyn azure_core::credentials::TokenCredential>>>,
+);
+
+impl std::fmt::Debug for AzureCredentialCache {
+	fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+		f.write_str("AzureCredentialCache")
+	}
 }
 
 #[apply(schema!)]
